requirements.txt
typer==0.12.3
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1
pyloudnorm==0.1.1
webrtcvad-wheels  # Windows互換wheel（Linuxは webrtcvad でも可）
tqdm==4.66.2
pyyaml==6.0.1
requests==2.31.0
# faiss-cpu==1.7.4  # Linux/Colab/RunPod推奨。Windowsは公式wheelが無い場合あり。必要時のみ個別インストール
